Restart from where stopped + show progress while running
If the program stops for some reason, the next run will continue its seach on franklin.genoox.com from where it stopped, instead of starting from the beginning.

During the search on franklin.genoox.com, it shows the progress of the search with the percentage of completion of the task.